move shufflingUtils.ts to shuffling.ts
